# openapi.yaml
openapi: 3.0.0
info:
    title: 'Petrogassa API'
    description: >
        API de autenticación con NestJS (modelo cerrado, sin registro público):
        un admin da de alta usuarios por invitación; cada invitado activa su
        cuenta definiendo nombre, apellido y contraseña. Login con JWT en cookies
        HttpOnly, refresh tokens rotados (hasheados en DB, con detección de
        reutilización) y recuperación de contraseña. Todas las respuestas usan un
        sobre estándar (ver esquemas `SuccessEnvelope` / `ErrorResponse`).
    version: '0.0.1'
servers:
    - url: /api
      description: Servidor local
...
paths:
...
    /certifications:
        get:
            tags: [Certifications]
            summary: Listar certificaciones (público)
            responses:
                '200':
                    description: Certificaciones obtenidas.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items:
                                                  $ref: '#/components/schemas/CertificationEntity'
        post:
            tags: [Certifications]
            summary: Crear certificación (admin)
            security: [{ cookieAuth: [] }]
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateCertificationDto'
            responses:
                '201': { description: Certificación creada. }
                '400': { $ref: '#/components/responses/ValidationError' }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
    /certifications/{id}:
        patch:
            tags: [Certifications]
            summary: Actualizar certificación (admin)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Certificación actualizada. }
                '404': { $ref: '#/components/responses/NotFound' }
        delete:
            tags: [Certifications]
            summary: Enviar certificación a la papelera (admin, soft delete)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Certificación movida a la papelera. }
                '404': { $ref: '#/components/responses/NotFound' }
    /certifications/admin/trash:
        get:
            tags: [Certifications]
            summary: Papelera de certificaciones (admin)
            security: [{ cookieAuth: [] }]
            responses:
                '200':
                    description: Certificaciones borradas (soft delete).
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data: { type: array, items: { $ref: '#/components/schemas/CertificationEntity' } }
                '401': { $ref: '#/components/responses/Unauthorized' }
    /certifications/{id}/restore:
        post:
            tags: [Certifications]
            summary: Restaurar una certificación de la papelera (admin)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Certificación restaurada. }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { description: La certificación no está en la papelera. }
    /certifications/{id}/permanent:
        delete:
            tags: [Certifications]
            summary: Eliminar una certificación definitivamente (admin; borra también el logo y el PDF del almacenamiento)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Certificación eliminada definitivamente. }
                '404': { $ref: '#/components/responses/NotFound' }
                '409': { description: 'El registro no está en la papelera: hay que enviarlo ahí antes de eliminarlo definitivamente.' }
...
    /change-log:
        get:
            tags: [ChangeLog]
            summary: Listar el registro de cambios (admin o auditor)
            description: >-
                Cada asiento dice qué se tocó, quién y cuándo, y bajo qué
                revisión del sitio. Los `details` traen el campo a campo: cuando
                viene `summary` es porque no se guardan los valores (campos
                largos y archivos).
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
                - { in: query, name: limit, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
                - { in: query, name: from, schema: { type: string, format: date }, description: Desde esta fecha inclusive. }
                - { in: query, name: to, schema: { type: string, format: date }, description: Hasta esta fecha inclusive. }
                - { in: query, name: entityType, schema: { type: string }, description: 'servicio, certificacion, cliente, novedad, titulo, configuracion, imagen-del-sitio' }
                - { in: query, name: action, schema: { type: string, enum: [creacion, modificacion, baja, restauracion, eliminacion] } }
            responses:
                '200':
                    description: Registro paginado, del más reciente al más viejo.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data:
                                              type: array
                                              items: { $ref: '#/components/schemas/ChangeLogEntry' }
                '400': { $ref: '#/components/responses/ValidationError' }
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
    /change-log/export:
        get:
            tags: [ChangeLog]
            summary: Exportar el registro a Excel (admin o auditor)
            description: >-
                Devuelve un .xlsx con una fila por campo cambiado. Mismos
                filtros que el listado, sin paginar. El nombre del archivo viaja
                en el Content-Disposition.
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: query, name: from, schema: { type: string, format: date } }
                - { in: query, name: to, schema: { type: string, format: date } }
                - { in: query, name: entityType, schema: { type: string } }
                - { in: query, name: action, schema: { type: string, enum: [creacion, modificacion, baja, restauracion, eliminacion] } }
            responses:
                '200':
                    description: La planilla.
                    content:
                        application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
                            schema: { type: string, format: binary }
                '400':
                    description: >-
                        El rango seleccionado supera las 50.000 filas
                        exportables; hay que acotarlo.
                '401': { $ref: '#/components/responses/Unauthorized' }
                '403': { $ref: '#/components/responses/Forbidden' }
    /contact:
        post:
            tags: [Contact]
            summary: Enviar mensaje de contacto (público, 5/min por IP)
            description: >
                Persiste el mensaje SIEMPRE (queda en el panel) y notifica por
                email al buzón de la empresa de forma best-effort.
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreateContactMessageDto'
            responses:
                '201':
                    description: Mensaje enviado.
                    content:
                        application/json:
                            schema:
                                allOf:
                                    - $ref: '#/components/schemas/SuccessEnvelope'
                                    - type: object
                                      properties:
                                          data:
                                              type: object
                                              properties:
                                                  id: { type: string, format: uuid }
                '400': { $ref: '#/components/responses/ValidationError' }
                '429': { $ref: '#/components/responses/TooManyRequests' }
    /contact/messages:
        get:
            tags: [Contact]
            summary: Listar mensajes de contacto (admin, paginado)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: query, name: status, schema: { type: string, enum: [nueva, leida] } }
                - { in: query, name: page, schema: { type: integer } }
                - { in: query, name: limit, schema: { type: integer } }
            responses:
                '200': { description: Mensajes obtenidos. }
                '401': { $ref: '#/components/responses/Unauthorized' }
    /contact/messages/{id}:
        patch:
            tags: [Contact]
            summary: Actualizar estado de un mensaje (admin)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Mensaje actualizado. }
                '404': { $ref: '#/components/responses/NotFound' }
        delete:
            tags: [Contact]
            summary: Eliminar mensaje (admin)
            security: [{ cookieAuth: [] }]
            parameters:
                - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
            responses:
                '200': { description: Mensaje eliminado. }
                '404': { $ref: '#/components/responses/NotFound' }
...
components:
    securitySchemes:
        cookieAuth:
            type: apiKey
            in: cookie
            name: accessToken
    responses:
        ValidationError:
            description: Error de validación (Bad Request)
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Unauthorized:
            description: No autenticado / credenciales o token inválidos
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        Forbidden:
            description: Autenticado pero sin permisos suficientes
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
        NotFound:
            description: El recurso solicitado no existe
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
...
        TooManyRequests:
            description: >
                Límite de peticiones excedido (global: 100/min por IP; endpoints
                de auth: 10/min; forgot-password: 3/min; contacto: 5/min;
                postulaciones: 20/min).
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/ErrorResponse'
    schemas:
        SuccessEnvelope:
            type: object
            description: Sobre estándar de respuestas exitosas.
            properties:
                success:
                    type: boolean
                    example: true
                statusCode:
                    type: integer
                    example: 200
                message:
                    type: string
                    example: 'Operación exitosa'
                data:
                    description: Carga útil (objeto, array o null).
                meta:
                    $ref: '#/components/schemas/PaginationMeta'
            required: [success, statusCode, message, data]
        PaginationMeta:
            type: object
            description: Metadata de paginación (presente solo en listas paginadas).
            properties:
                totalItems: { type: integer, example: 145 }
                itemsPerPage: { type: integer, example: 10 }
                currentPage: { type: integer, example: 2 }
                totalPages: { type: integer, example: 15 }
                hasNextPage: { type: boolean, example: true }
                hasPrevPage: { type: boolean, example: true }
                nextPageUrl:
                    type: string
                    nullable: true
                    example: '/api/products?limit=10&page=3'
                prevPageUrl:
                    type: string
                    nullable: true
                    example: '/api/products?limit=10&page=1'
        ErrorResponse:
            type: object
            description: Sobre estándar de errores (lo formatean los Exception Filters).
            properties:
                success: { type: boolean, example: false }
                statusCode: { type: integer, example: 400 }
                message:
                    type: string
                    example: 'La contraseña debe tener al menos 8 caracteres...'
                errors:
                    type: array
                    nullable: true
                    items: { type: string }
                timestamp: { type: string, format: date-time }
                path: { type: string, example: '/api/auth/activate' }
...
        CertificationEntity:
            type: object
            properties:
                id: { type: string, format: uuid }
                title: { type: string, example: 'ISO 9001:2015' }
                subtitle: { type: string, example: 'Calidad' }
                description: { type: string }
                isFeatured: { type: boolean }
                logoImage: { type: string, nullable: true, description: URL absoluta }
                certificatePdf:
                    type: string
                    nullable: true
                    description: URL absoluta pública del certificado en PDF (null si no se cargó). El sitio muestra "Ver / Descargar certificado" cuando no es null.
                sortOrder: { type: integer }
                deletedAt: { type: string, format: date-time, nullable: true, description: 'En la papelera si no es null' }
        CreateCertificationDto:
            type: object
            required: [title, subtitle, description]
            properties:
                title: { type: string, maxLength: 120 }
                subtitle: { type: string, maxLength: 120 }
                description: { type: string, maxLength: 2000 }
                isFeatured: { type: boolean }
                logoImage: { type: string, description: 'Key de media (media/AAAA/MM/uuid.ext). Se acepta también la URL absoluta que devuelve el GET: se normaliza a key. Cualquier otro valor da 400.' }
                certificatePdf: { type: string, description: 'Key de media (media/AAAA/MM/uuid.ext). Se acepta también la URL absoluta que devuelve el GET: se normaliza a key. Cualquier otro valor da 400.' }
                sortOrder: { type: integer }
...
        CreateContactMessageDto:
            type: object
            required: [name, email, message]
            properties:
                name: { type: string, maxLength: 100 }
                email: { type: string, format: email }
                phone: { type: string, maxLength: 30 }
                subject: { type: string, maxLength: 150 }
                message: { type: string, minLength: 10, maxLength: 5000 }
                referencia:
                    type: string
                    maxLength: 200
                    description: >-
                        Campo TRAMPA (honeypot). El formulario debe incluirlo
                        oculto (fuera de pantalla, tabindex="-1",
                        autocomplete="off") y dejarlo SIEMPRE vacío: si llega
                        con contenido, el mensaje se descarta en silencio y la
                        respuesta es 201 igual, para no avisarle al bot.
        ChangeLogEntry:
            type: object
            description: Un asiento del registro de cambios. Append-only.
            properties:
                id: { type: string, format: uuid }
                occurredAt: { type: string, format: date-time }
                revision: { type: string, example: '03', description: 'Revisión del sitio bajo la que cayó el cambio. Dos dígitos, sin punto; a los tres dígitos vuelve a 01 y la fecha es lo que desambigua.' }
                entityType: { type: string, example: certificacion }
                entityLabel:
                    type: string
                    example: 'ISO 9001:2015'
                    description: >-
                        Cómo se llamaba el registro CUANDO se lo tocó. Congelado
                        a propósito: si después se renombra o se elimina, el
                        asiento sigue diciendo de qué se trataba.
                action: { type: string, enum: [creacion, modificacion, baja, restauracion, eliminacion] }
                actorLabel:
                    type: string
                    example: 'Nicolás Espinoza <nicolas.espinoza@petrogassa.com>'
                    description: Quién lo hizo. Sobrevive al borrado de esa cuenta.
                details:
                    type: array
                    items:
                        type: object
                        properties:
                            fieldLabel: { type: string, example: Descripción }
                            previousValue: { type: string, nullable: true }
                            newValue: { type: string, nullable: true }
                            summary:
                                type: string
                                nullable: true
                                description: >-
                                    Cuando viene, reemplaza a los valores: pasa
                                    con los campos largos (solo se registra que
                                    cambiaron) y con los archivos.
                            valueKind: { type: string, enum: [texto, booleano, numero, fecha, archivo, lista] }
...